+++ R/md_stock1_163.R
@@ -25,7 +25,7 @@ md_stock_spotall_163 = function(symbol = c('a','index'), only_symbol = FALSE, sh
         date = data_date, #as.Date(substr(jsonDat$time,1,10)), 
         time = jsonDat$time#,
         #strptime(jsonDat$time, '%Y-%m-%d %H:%M:%S', tz = 'Asia/Shanghai')
-      )][, .(symbol, name, date, open, high, low, close=price, prev_close=yestclose, change=updown, change_pct=percent*100, volume, amount=turnover, turnover=hs*100, cap_market=mcap, cap_total=tcap, pe_last=pe, eps=mfsum, net_income, revenue, plate_ids, time=as.POSIXct(time))
+      )][, .(symbol, name, date, open, high, low, close=price, close_prev=yestclose, change=updown, change_pct=percent*100, volume, amount=turnover, turnover=hs*100, cap_market=mcap, cap_total=tcap, pe_last=pe, eps=mfsum, net_income, revenue, plate_ids, time=as.POSIXct(time))
       ][,`:=`(
         province = sub('.*(dy\\d+).*', '\\1', plate_ids),
         plate_ids = sub('dy\\d+','',plate_ids)
@@ -39,7 +39,7 @@ md_stock_spotall_163 = function(symbol = c('a','index'), only_symbol = FALSE, sh
       
       jsonDF = setDT(jsonDF)[,`:=`(
         date = data_date#as.Date(substr(jsonDat$time,1,10))
-      )][, .(symbol, name, date, open, high, low, close=price, prev_close=yestclose, change=updown, change_pct=percent*100, volume=volume/100, amount=turnover, time=as.POSIXct(time))]
+      )][, .(symbol, name, date, open, high, low, close=price, close_prev=yestclose, change=updown, change_pct=percent*100, volume=volume/100, amount=turnover, time=as.POSIXct(time))]
     }
     
     return(jsonDF[, `:=`(market = mkt, region = 'cn')])
@@ -94,7 +94,7 @@ md_stock_spotall_163 = function(symbol = c('a','index'), only_symbol = FALSE, sh
   
   df = df_stock_cn[,unit := 'CNY'][, symbol := check_symbol_for_yahoo(symbol, market)]#[, mkt := NULL][]
   
-  cols_rm = intersect(names(df), c('sector', 'industry', 'province', 'plate_ids', 'region')) # , 'prev_close', 
+  cols_rm = intersect(names(df), c('sector', 'industry', 'province', 'plate_ids', 'region')) # , 'close_prev', 
   if (length(cols_rm)>0) df = df[, (cols_rm) := NULL]
   return(df)
 }
@@ -105,7 +105,7 @@ md_stock_spotall_163 = function(symbol = c('a','index'), only_symbol = FALSE, sh
 # hq.sinajs.cn/list=sz150206
 # http://qt.gtimg.cn/q=sz000001
 md_stock_spot_tx = function(symbol1, only_syb_nam = FALSE, ...) {
-  V1=V2=doc=.=symbol=name=high=low=prev_close=change=change_pct=volume=amount=turnover=cap_market=cap_total=pb=pe_last=pe_trailing=pe_forward=NULL
+  V1=V2=doc=.=symbol=name=high=low=close_prev=change=change_pct=volume=amount=turnover=cap_market=cap_total=pb=pe_last=pe_trailing=pe_forward=NULL
   
   syb = paste0(check_symbol_for_tx(symbol1), collapse=',')
   dt = try(readLines(sprintf('http://qt.gtimg.cn/q=%s', syb)), silent = TRUE)
@@ -122,7 +122,7 @@ md_stock_spot_tx = function(symbol1, only_syb_nam = FALSE, ...) {
     return(dt[,.(symbol = check_symbol_for_yahoo(symbol1), name = V1)])
   }
   
-  colnames_en = c('name', 'symbol', 'close', 'prev_close', 'open',
+  colnames_en = c('name', 'symbol', 'close', 'close_prev', 'open',
                   'volume', 'buy', 'sell', 
                   'bid1', 'bid1_volume', 'bid2', 'bid2_volume', 'bid3', 'bid3_volume', 'bid4', 'bid4_volume', 'bid5', 'bid5_volume',
                   'ask1', 'ask1_volume', 'ask2', 'ask2_volume', 'ask3', 'ask3_volume', 'ask4', 'ask4_volume', 'ask5', 'ask5_volume',
@@ -137,10 +137,10 @@ md_stock_spot_tx = function(symbol1, only_syb_nam = FALSE, ...) {
   setnames(dt, colnames_en)
   
   num_cols = c(
-    'open', 'high', 'low', 'close', 'prev_close', 'change', 'change_pct', 'volume', 'amount', 'turnover', 'cap_market', 'cap_total', 'pb', 'pe_last', 'pe_trailing', 'pe_forward'
+    'open', 'high', 'low', 'close', 'close_prev', 'change', 'change_pct', 'volume', 'amount', 'turnover', 'cap_market', 'cap_total', 'pb', 'pe_last', 'pe_trailing', 'pe_forward'
   )
   dt = dt[,.(
-    symbol, name, date, open, high, low, close, prev_close, change, change_pct, volume, amount, turnover, cap_market, cap_total, pb, pe_last, pe_trailing, pe_forward#, 
+    symbol, name, date, open, high, low, close, close_prev, change, change_pct, volume, amount, turnover, cap_market, cap_total, pb, pe_last, pe_trailing, pe_forward#, 
     #buy, sell, 
     #bid1, bid1_volume, bid2, bid2_volume, bid3, bid3_volume, bid4, bid4_volume, bid5, bid5_volume, 
     #ask1, ask1_volume, ask2, ask2_volume, ask3, ask3_volume, ask4, ask4_volume, ask5, ask5_volume
@@ -161,6 +161,127 @@ md_stock_spot_tx = function(symbol1, only_syb_nam = FALSE, ...) {
   return(dt[, `:=`(unit = 'CNY')])
 }
 
+
+
+md_stock_info1_163 = function(symbol1, str_income_hist = FALSE, ...) {
+  . = N = V1 = X1 = X2 = X3 = X4 = income_by = mkt = rd = rid = NULL
+  
+  # symbol1 = '600036'
+  # http://quotes.money.163.com/f10/gszl_600036.html#01f01
+  # http://quotes.money.163.com/service/gszl_600036.html?type=cp
+  # http://quotes.money.163.com/service/gszl_600036.html?type=hy
+  # http://quotes.money.163.com/service/gszl_600036.html?type=dy
+  
+  # skip index
+  if (check_symbol_cn(symbol1)[, mkt == 'index' || is.na(mkt)]) return(NULL)
+  # symbol1 = '000001'
+  stk_price = md_stock_spot_tx(symbol1, only_syb_nam=TRUE)
+  # columns
+  income_strcols = c('income', 'cost', 'profit', 'gross_margin', 'profit_proportion')
+  employee_strcols = c('number', 'proportion')
+  
+  if (str_income_hist) {
+    income_hist_preprocess = function(dat, type) {
+      income_strcols = c('income', 'cost', 'profit', 'gross_margin', 'profit_proportion')
+      
+      rbindlist(lapply(
+        split(income_product[
+          , rid := 0
+        ][V1 == income_product[1,1], rid := 1
+        ][, rid := cumsum(rid)][], by = 'rid'), 
+        function(x) {
+          setnames(
+            x[-c(1:2),
+            ][, report_date := x[1,5]
+            ][, income_by := type
+            ][, c(8,9,1:6)], 
+            c('report_date', 'income_by', 'variable', income_strcols)
+          )
+        }
+      ))[, (income_strcols) := lapply(.SD, function(x) {
+        as.numeric(gsub(',|\\s|-|%', '', x))
+      }), .SDcols = income_strcols][]
+    }
+    
+    income_product = load_read_csv(
+      sprintf('http://quotes.money.163.com/service/gszl_%s.html?type=cp', symbol1), 
+      encode = 'GBK', csv_header = FALSE
+    )
+    income_product2 = income_hist_preprocess(income_product, type = 'product')
+    
+    
+    income_industry = load_read_csv(
+      sprintf('http://quotes.money.163.com/service/gszl_%s.html?type=hy', symbol1), 
+      encode = 'GBK', csv_header = FALSE
+    )
+    income_industry2 = income_hist_preprocess(income_industry, type = 'industry')
+    
+    
+    income_location = load_read_csv(
+      sprintf('http://quotes.money.163.com/service/gszl_%s.html?type=dy', symbol1), 
+      encode = 'GBK', csv_header = FALSE
+    )
+    income_location2 = income_hist_preprocess(income_location, type = 'location')
+    
+    
+    retdat = list(str_income = cbind(
+      stk_price,
+      rbind(income_product2, income_industry2, income_location2)
+    ))
+  } else {
+    # load all tables
+    datweb = read_html(sprintf(
+      'http://quotes.money.163.com/f10/gszl_%s.html#01f01', 
+      sub("^.*?([0-9]+).*$",'\\1', symbol1)
+    ))
+    report_date = 
+      data.table(
+        rd = html_text(html_nodes(datweb, 'div.report_date span'))
+      )[, report_date := sub('.*([0-9]{4}-[0-9]{2}-[0-9]{2}).*', '\\1', rd) 
+      ][,.N, by = report_date][order(-N)][1, .(report_date)] 
+    
+    tbls = rvest::html_table(datweb, fill = TRUE, header = FALSE)
+    
+    profile = 
+      rbind(
+        setDT(tbls[[4]])[1:9,   .(variable = X1, value = X2)], 
+        setDT(tbls[[4]])[1:9,   .(variable = X3, value = X4)], 
+        setDT(tbls[[4]])[10:12, .(variable = X1, value = X2)]
+      )
+    
+    ipo = setDT(tbls[[5]])[, .(variable = X1, value = X2)][]
+    info = cbind(stk_price, rbindlist(list(profile=profile, ipo=ipo), idcol = 'info_type'))
+    
+    
+    str_income = tbls[7:9]
+    names(str_income) = c('prodcut', 'industry', 'location')
+    str_income = rbindlist(
+      lapply( str_income, function(x) setnames(x[-1,], c('variable', income_strcols)) ), 
+      idcol = 'income_by'
+    )[, (income_strcols) := lapply(.SD, function(x) {
+      as.numeric(gsub(',|\\s|-|%', '', x))
+    }), .SDcols = income_strcols][]
+    str_income = cbind(stk_price, report_date, str_income)
+    
+    str_employee = tbls[10:11]
+    names(str_employee) = c('education', 'work_type')
+    str_employee = rbindlist(
+      lapply(str_employee, function(x) setnames(x[-1,], c('variable', employee_strcols))), 
+      idcol = 'employee_by'
+    )[, (employee_strcols) := lapply(.SD, function(x) {
+      as.numeric(gsub('\\s|%', '', x))
+    }), .SDcols = employee_strcols]
+    str_employee = cbind(stk_price, report_date, str_employee)
+    
+    retdat = list(
+      info = info, 
+      str_income = str_income, 
+      str_employee = str_employee
+    )
+  }
+  
+  return(retdat)
+}
 
 #' @import data.table
 #' @importFrom readr read_csv locale col_date col_character col_double col_integer
@@ -197,7 +318,7 @@ md_stock_hist1_163 = function(symbol1, from='1900-01-01', to=Sys.Date(), zero_rm
   # dt <- load_read_csv(link, 'GBK')
   
   # set names of datatable
-  cols_name = c('date', 'symbol', 'name', 'open', 'high', 'low', 'close', 'prev_close', 'change', 'change_pct', 'volume', 'amount', 'turnover', 'cap_market', 'cap_total')
+  cols_name = c('date', 'symbol', 'name', 'open', 'high', 'low', 'close', 'close_prev', 'change', 'change_pct', 'volume', 'amount', 'turnover', 'cap_market', 'cap_total')
   setnames(dt, cols_name)
   setkeyv(dt, 'date')
   
@@ -314,7 +435,7 @@ md_stock_divsplit1_163 = function(symbol1, from=NULL, to=NULL, ret = c('div', 's
     rvest::html_table(., fill = TRUE, header = TRUE)
   
   
-  tbl_divspl = setDT(tbls[[4]])[-1][, lapply(.SD, function(x) replace(x, x=='--', NA))][,c(3:7,8)]
+  tbl_divspl = setDT(tbls[[4]])[-1][, lapply(.SD, function(x) replace(x, x=='--', NA))][,c(3:8)]
   setnames(tbl_divspl, c('songgu', 'zhuanzeng', 'fenhong', 'date0', 'date1', 'date2'))
   if (length(unique(unlist(tbl_divspl))) == 1 & nrow(tbl_divspl) == 1) {
     tbl_divspl = tbl_divspl[.0]
@@ -334,7 +455,7 @@ md_stock_divsplit1_163 = function(symbol1, from=NULL, to=NULL, ret = c('div', 's
         )[.0]
     } else {
       div_spl[['div']] = tbl_divident[,.(
-          date      = date0,
+          date      = date1,
           dividends = fenhong/10
         )]
     }
@@ -342,8 +463,8 @@ md_stock_divsplit1_163 = function(symbol1, from=NULL, to=NULL, ret = c('div', 's
   
   if ('spl' %in% ret) { # split
     tbl_split = rbind(
-      tbl_divspl[, .(spl = songgu, date = date0)][spl > 0],
-      tbl_divspl[, .(spl = zhuanzeng, date = date0)][spl > 0]
+      tbl_divspl[, .(spl = songgu, date = date1)][spl > 0],
+      tbl_divspl[, .(spl = zhuanzeng, date = date1)][spl > 0]
     )[, lapply(.SD, sum), keyby = date]
     
     if (nrow(tbl_split) == 0) {
@@ -415,17 +536,19 @@ md_stock_163 = function(symbol, from='1900-01-01', to=Sys.Date(), print_step=1L,
   } else if (type == 'adjfactor') (
     dat_list = load_dat_loop(symbol, 'md_stock_divsplit1_163', args = list(from = from, to = to), print_step=print_step)
     
-  )
+  ) else if (type == 'info') {
+    dat_list = load_dat_loop(symbol, 'md_stock_info1_163', args = list(...), print_step=print_step)
+  }
   
   rmcols_func = function(x) {
     name = NULL
     
-    cols_rm = intersect(names(x), c('change')) #,'prev_close', 'change_pct'
+    cols_rm = intersect(names(x), c('change')) #,'close_prev', 'change_pct'
     if (length(cols_rm)>0) x = x[, (cols_rm) := NULL]
     if ('name' %in% names(x)) x = x[, name := gsub('\\s', '', name)]
     return(x)
   }
-  if (type != 'adjfactor') {
+  if (type %in% c('history', 'spot')) {
     if (inherits(dat_list, 'list')) {
       dat_list = lapply(dat_list, rmcols_func)
     } else if (inherits(dat_list, 'data.frame')) {
@@ -463,7 +586,7 @@ md_fund_spotall_163 = function() {
   dat_df = rbindlist(dat_lst, idcol = 'fund')
   setnames(dat_df, tolower(names(dat_df)))
   
-  dat_df = dat_df[, .(symbol = check_symbol_for_yahoo(symbol), date = md_stock_spot_tx('^000001')$date, name, open, high, low, close = price, prev_close = yestclose, change_pct = percent*100, volume, amount = turnover, market = fund, unit = 'CNY')] # change=updown, sname, 
+  dat_df = dat_df[, .(symbol = check_symbol_for_yahoo(symbol), date = md_stock_spot_tx('^000001')$date, name, open, high, low, close = price, close_prev = yestclose, change_pct = percent*100, volume, amount = turnover, market = fund, unit = 'CNY')] # change=updown, sname, 
   
   return(dat_df)
 }